requirements.txt
pandas
numpy
scikit-learn
xgboost
pyyaml
joblib
mlflow
dagshub
python-dotenv
dvc
flask